Revert "Build smooth-filter"
