+++ Makefile
@@ -4,7 +4,7 @@ LIB = ./lib
 LIB_TEST = $(LIB)/test.lua
 LIB_NEBLUA = $(LIB)/neblua.lua
 
-build: build-psd build-lcd-pixel-filter build-icon-and-cursor build-smooth-filter
+build: build-psd build-lcd-pixel-filter build-icon-and-cursor
 
 build-psd: $(DIST) $(LIB_NEBLUA)
 	DIST_DIR=$(DIST) $(LUA) ./psd/build.lua
@@ -14,9 +14,6 @@ build-lcd-pixel-filter: $(DIST) $(LIB_NEBLUA)
 
 build-icon-and-cursor: $(DIST) $(LIB_NEBLUA)
 	DIST_DIR=$(DIST) $(LUA) ./icon-and-cursor/build.lua
-
-build-smooth-filter: $(DIST) $(LIB_NEBLUA)
-	DIST_DIR=$(DIST) $(LUA) ./smooth-filter/build.lua
 
 prepare: $(LIB_TEST) $(LIB_NEBLUA)
