+++ readme.md
@@ -2,26 +2,33 @@
 
 ## Scripts
 
-- [**Export as PSD**](psd/readme.md) Exports sprite as Adobe Photoshop file format
-- [**Export as ico, cur, ani**](icon-and-cursor) Exports sprite as Windows static/animated cursor file.
-- [**LCD Pixel Filter**](lcd-pixel-filter/readme.md) Filter effect that looks like an LCD screen.
-- [Smooth Filter ](smoothFilter/readme.md) Smooghing filter effect.
+- [Export as PSD ](psd/readme.md)
+  Applies export as photoshop file format (.psd)
+- [LCD Pixel Filter](lcd-pixel-filter/readme.md)
+  Filter looks like LCD display
+- [Smooth Filter ](smoothFilter/readme.md)
+  Filter smooths pixel-art
 
 ## Script Installation
 
 Aseprite scripts are provided as `.lua` file. You have to place then into script folder.
 
-1. Open scripts folder (via `File` -> `Scripts` -> `Open Scripts Folder`)
-2. Place `.lua` file that you have downloaded.
-3. Reload script folder or restart Aseprite.
-4. You can find scripts in `File` -> `Scripts` menu.
+1. Open scripts folder
+   (In Aseprite's menu, File -> Scripts -> Open Scripts Folder)
+2. Place `*.lua` file you've downloaded.
+3. Reload script folder
+   (Via File -> Scripts -> Rescan Scripts Folder, or just restart Aseprite)
+
+## Running Script
+
+In Aseprite's menu, you should find scripts you've installed in File -> Scripts
 
 ## License
 
-[MIT License](https://github.com/Tsukina-7mochi/aseprite-scripts/blob/master/LICENSE)
+- All script in this repo are provided with [MIT License](https://github.com/Tsukina-7mochi/aseprite-scripts/blob/master/LICENSE)
 
 ## Contact
 
 - GitHub issues
-- Mentions / direct messages in X (formarly Twitter): [@Tsukina_7mochi](https://twitter.com/Tsukina_7mochi)
+- Twitter's direct message [@Tsukina_7mochi](https://twitter.com/Tsukina_7mochi)
