--- smooth-filter/build.lua
@@ -1,12 +1,0 @@
-local neblua = require("lib.neblua")
-
-local distDir = os.getenv("DIST_DIR")
-
-neblua.bundle {
-    entry = "main",
-    output = distDir .. "/Smooth Filter.lua",
-    include = {
-        "./main.lua",
-    },
-    rootDir = "./smooth-filter",
-}
